requirements.txt
pandas
pyarrow
requests
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-core
python-json-logger
python-dotenv
typing-extensions
fastapi